refactor(errors): move eftype
